fixed mentions being allowed to be prefixes

+++ slash/changeprefix.js
@@ -1,5 +1,6 @@
 const prefixSchema = require("../Schemas/prefixSchema");
 const { newCache } = require("../Utils/prefix-load");
+const { MessageMentions } = require("discord.js");
 module.exports = {
   name: "changeprefix",
   description: "Changes The Prefix",
@@ -8,6 +9,13 @@ module.exports = {
   ],
   run: async (client, interaction) => {
     const prefix = interaction.options.getString("prefix");
+    if (
+      MessageMentions.USERS_PATTERN.test(prefix) ||
+      MessageMentions.ROLES_PATTERN.test(prefix) ||
+      MessageMentions.CHANNELS_PATTERN.test(prefix)
+    ) {
+      return interaction.reply("Mentions are not allowed to be prefixes");
+    }
     if (interaction.member.permissions.has("MANAGE_GUILD")) {
       if (prefix === "b!") {
         await prefixSchema.deleteOne({ _id: interaction.guild.id });
